Extract visualiser to its own component file for testing purposes
